Introduce Profiler utility class
