+++ web/modules/user_types/organizations/src/Entity/Organization.php
@@ -4,6 +4,7 @@ namespace Drupal\organizations\Entity;
 
 use Drupal\Core\Session\AccountInterface;
 use Drupal\user_bundle\Entity\TypedUser;
+use Drupal\user_types\Utility\Profiler;
 
 class Organization extends TypedUser {
 
@@ -16,22 +17,11 @@ class Organization extends TypedUser {
   }
 
   public function isManager(AccountInterface|int $account) {
-    return $this->getManager()->id() == $this->getUid($account);
+    return $this->getManager()->id() == Profiler::id($account);
   }
 
   public function isOwnerOrManager(AccountInterface|int $account) {
-    return $this->id() == $this->getUid($account) || $this->isManager($account);
+    return $this->id() == Profiler::id($account) || $this->isManager($account);
   }
 
-  /**
-   * Helper to get uid of an account.
-   *
-   * @param \Drupal\Core\Session\AccountInterface|int $account
-   *   The account or the uid.
-   * @return \Drupal\Core\Session\AccountInterface|int
-   *   The uid.
-   */
-  private function getUid(AccountInterface|int $account) {
-    return $account instanceof AccountInterface ? $account->id() : $account;
-  }
 }
