+++ web/modules/youvo/src/Utility/FieldAccess.php
@@ -38,39 +38,4 @@ abstract class FieldAccess {
       $field->getName() : $field;
   }
 
-  /**
-   * With different authorization methods the account object may be a
-   * AccountProxy or a TokenAuthUser. Use this helper to determine whether
-   * the account is of a specific user type.
-   */
-  protected static function isUserType(
-    AccountInterface $account,
-    string $type,
-    string $class
-  ) {
-    if ($account instanceof AccountProxyInterface) {
-      $account = $account->getAccount();
-      if (class_exists('Drupal\\simple_oauth\\Authentication\\TokenAuthUser') &&
-        $account instanceof \Drupal\simple_oauth\Authentication\TokenAuthUser) {
-        return $account->bundle() == $type;
-      }
-      return $account instanceof $class;
-    }
-    return $account instanceof $class;
-  }
-
-  /**
-   * Helper to determine if account is creative.
-   */
-  protected static function isCreative(AccountInterface $account) {
-    return self::isUserType($account, 'user', 'Drupal\\creatives\\Entity\\Creative');
-  }
-
-  /**
-   * Helper to determine if account is organization.
-   */
-  protected static function isOrganization(AccountInterface $account) {
-    return self::isUserType($account, 'organization', 'Drupal\\organizations\\Entity\\Organization');
-  }
-
 }
